+++ examples/data-sources/nxos_isis_vrf/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_isis_vrf" "example" {
+  instance_name = "ISIS1"
+  name          = "default"
+}