+++ examples/resources/nxos_isis_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_isis_vrf.example "sys/isis/inst-[ISIS1]/dom-[default]"